Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseDataModel  1e3f1730abb03cdaf4a1ae4644672b025608f743
EntityId rewrite

This includes

* Dropping of id prefix configurability (as previously discussed)
* Separation of the id class and the entity id DataValue (EntityId
no longer derives from DataValue. A new EntityIdDataValue
composites in a EntrityId)
* The EntityId class now has as fields entity type (string) and
id serialization (string)
* New ItemId and PropertyId derivatives, which in the constructor
only take the id serialization. The long term goal is to make
EntityId abstract
* Serialization format compatibility where required. The format
of the DataValue will remain the same for now

Some implications

* No more global state is used
* No more semi-dependency from DataModel on lib
* getPrefixedId is no longer evil and thus not deprecated any more
* EntityId and derivatives no longer expose the id format. A
FooIdDerivative that has a md5 in it can be created if that'd make sense
* One can type hint against ItemId or PropertyId, which are
guaranteed to be valid ids for the correct entity type

This is a breaking change for Wikibase, which needs to have
https://gerrit.wikimedia.org/r/#/c/80506/ merged soon after
this commit goes in.

Change-Id: I9b05e8305967f7cd3cf121bcd28d59e820c3c5b0