Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  9dff77ea18c1d1ccd5077278c8c1f9ecc178aefc
Enforce wgSecureLogin for login pages

If wgSecureLogin is set, redirect to https when accessing login pages.
This shouldn't be needed in most cases, but in case a wiki without
wgSecureLogin is using SUL, then this will ensure communication with
the central wiki goes over https if wfCanIPUseHTTPS is also true.

Change-Id: Ia1232ac39e8054748b0304d17efc88c4dc1c7c71