Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  051867c3ab3dbf8074960a756fc1051e6ebb18ee
Actually use GeneratedContentNode in AlienNode

ce.AlienNode was mixing in GeneratedContentNode, but wasn't actually using
it properly. This commit fixes this so that AlienNode can benefit from
the <meta>/<link>/<style> stripping code in GeneratedContentNode.

Also added a getHashObject() function to dm.AlienNode to summarize the
domElements attribute because ve.getHash() chokes on it otherwise.

Change-Id: Ief3be94f9730297abe0e3c57506b81a8ff1d136d