Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  0d6c4aaf4e3621d576c7a15a84b243af71af793b
Remove hack from tree builder to not foster metas.

Various passes still rely on metas not being hoisted out of tables. In
order to preserve their location, typeof=mw:* metas are converted to
comments, which aren't fostered [1], and then switched back after the tree
is built.

The shadow metas for markTreeBuilderFixups are emitted as comments.

[1] According to the spec, http://www.whatwg.org/specs/web-apps/current-work/multipage/tree-construction.html#insert-a-comment

Bug: 53284
Change-Id: I443af96c02d59afc39c6c4f9477c4ffbc0495df4