Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  2e8a711c485739b4fc97d517954fd4e8524a1448
Config for site group to use for langlinks.

This allows the site group to show as language links to be
configurable, instead of always using the wiki's own group.
This is important for use cases like Wikimedia Commons.

(Note: this has unit tests, but I havn't tested it in the UI)

bug: 52788
Change-Id: I4e71671c8b617ba74f6d3b2b5734350e99d855c4