Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  8e6e07c75f3d0d858b459eb4c0aede4b4c831307
Handle backtrace where Lua can't guess the function name

Current logic is to display the funciton name if Lua provides us with
one, "main" if it's at the main level, or "?" if it's a C function or a
tail call. But we're not handling if it's a Lua call but Lua can't guess
a name for the function; use "?" for that too.

Change-Id: I938b5e5ca55cf4990dbcbb0db8dd8fc93b03bf15