Updated mediawiki/extensions Project: mediawiki/extensions/Flow  e95985d7572a9808cf72166dffcdc26a83e607de
Remove visualEditor.standalone RL dependency

I don't think there's an actual dependency at the moment; I can add
items on my local wiki without VE installed.

It would be nice if this was a configuration option, since VE is
somewhat hard to configure on a labs instance (but is easy in
mediawiki-vagrant). Though we should talk about whether a Flow that can
work without VE should be a deliverable.

Change-Id: I9b471a5708fbb446b843337c9073ba822ec326ae