Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  454537887af7f22975b043ecb1cc4e6348a43ffa
Using "poll" as positioning method for first file input

because CSS seems to be sometimes loaded later, leading a slightly wrong
position of the file input control.

Secondly, hiding file inputs after user chose uploading from Flickr
because it is on top of the Flickr URL input and when clicked, is
opening the file selection dialog.

Bug: 53522
Change-Id: I46ebbf82ef3916af5d8c1a117bde07ffd5149e87