Updated mediawiki/extensions Project: mediawiki/extensions/BetaFeatures  87d152e7c3ea7d495b80115da606f9ab8735a72c
Add convenience functions, fix auto-enroll

Adds a function that will check whether a user is enrolled in a
particular feature experiment.

Also changes the structure of the auto-enroll system so it works a
little more sanely. A little.

And introduced a concept of required fields for features...but not
using it yet, since I've erased the only required field from existence
since writing that bit.

Change-Id: Icfe6ad42d6d8c1a4c6ac00cf8bdbe9f8a0604a53