Updated mediawiki/extensions Project: mediawiki/extensions/DataValues  cacf5fd6e4a168ca8bd2bb135c4d3320827a96a7
Reorganize to new component structure.

This is the second step in implementing the reorganization outlined in
http://lists.wikimedia.org/pipermail/wikidata-tech/2013-August/000204.html

All changes are internal organization. The only thing that changed
from the outside is that the old component specific entry points
no longer exist. These did not have any users, so this ought to not
cause any problems.

The next and final step in the reorganization is to move the contents
of the DataValuesCommon, DataValuesInterfaces and ValueView directories
into their own git repos, and to promote the contents of the DataValues
directory to the root of the git repo. That change will obviously
require users to change some things :)

Change-Id: Ie8e50a244fe9a6c9b193182343c3784671d15022