Updated mediawiki/extensions Project: mediawiki/extensions/SocialProfile  cdd1608192822c2229d1d0998e4c1efcec2fa597
SocialProfile: removed UserStatus & fixed some brainfarts.

The UserStatus feature had been broken since the introduction of
ResourceLoader because it made extensive use of inline JS. A different
implementation of the feature will be committed later on (but it'll be a
separate extension).

Updated README and some other places to use HTTPS URLs instead of HTTP.

Fixed errors in few UserStats special pages, which were using the old
method for loading one CSS file instead of the ResourceLoader method for
loading a RL module.

Bumped version number to 1.6.1.

Change-Id: If78fae806a4644852b1c97f97b55c7bf208f72ab