Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  8dc157bc091ebb34911dfefd8bb4b483d2b7782d
Test gzdecode() in TemplateDataBlob::newFromDatabase()

gzdecode() has been introduced in PHP 5.4 although gzencode() has been
there for a while.  Wikimedia is still using PHP 5.3, so the
TemplateDataBlob::newFromDatabase() would trigger a fatal error whenever
it is passed compressed JSON :-(

The test does not fix the root cause (that needs a fallback function in
the TemplateData extension), but it does highlight the issue in PHP 5.3.x.

Updated a @return comment, correcting the class being returned.

bug: 54058
Change-Id: I8357b474165f4317982b9c924cf0afe68650d677