Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  b053e86d67eb73fb157ec31a4344b9fcd3399ac4
Clean up CentralNotice Translation Metadata

Right now the translation metadata job is running very slowly --
part of this problem is that there is a lot of junk metadata
that we have to process.

So... the maintenance script cleans things up. And I'm now putting
the banner id in with the revtag object so that we can eventually
look to see if the banner is archived or attached only to archived
campaigns or something.

E.g. this patch alone will take 800 objects on meta down to ~250.

Bug: 53769
Bug: 53792
Change-Id: If9ae70977ee0f008f65ef8f742acafb991b4d221