Updated mediawiki/extensions Project: mediawiki/extensions/BetaFeatures  4cd3567d833a78928d7e20d9268ba5b1cb3cd11f
Increment and decrement memcached on pref save

Now *that* is cool.

The cache now handles granular changes to the user count (and on a
pref-by-pref basis) using the $wgMemc->incr and ->decr methods. Also
had to mess about with the database and update functions, but they seem
to be working just fine still.

Change-Id: I2de2a2fd9d36232ec76725636fc3d2c2eb205bb6