Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  5d68f48e41ff9937ea7194ba09fdb52b4b5011cf
Cache Banner Message Field Definitions

This makes it possible to not have to wikitext parse the banner
content every time we load a banner. This is most useful for
translation when we have to run through each banner to figure
out what needs translating.

This also removes the post save job run configuration variable
because the job should be significantly faster to run now.

Bug: 53769
Bug: 53792
Bug: 53674
Change-Id: I82c8b0ede730dd5e06d796609a63475367e5a571