Updated mediawiki/extensions Project: mediawiki/extensions/MassMessage  8ef3176d2ebc7a9abc41a6d80e86c8c7cefe04cb
Only use wfWikiId() when submitting the job

Internally it makes more sense to use the actual dbname,
this makes it easier to use wgConf. Since we only need to
use the database prefix once, we should just change it in
that instance rather than having an inconsistent internal
structure (some with prefix, some without).

This reverts commit d688e4862b04ec3ff678dee9de8bd7f446ecbe2a.

Change-Id: I16c37bc1e0034f14ef2611ef769821e1b600676c