Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  807d95380f068a9afe5e0dbdce2d0258e468e8f6
Don't fail autologin for unattached accounts

For the post-login icons, we don't want to fail logging in to e.g. all
language wiktionaries if the user happens to not have an account on
en.wiktionary or the en.wiktionary account exists but isn't attached. So
don't require an attached account for the icon auto-logins.

But the script autologin is a different matter. In that case we *do*
want the local account to exist and be attached (that was bug 51603), so
if we get to that code path make it so.

Bug: 54292
Change-Id: I0ca1a33500784a30e9f04d1d1a5dd13901084441