Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  7566d051b0d5f80772a545927feea0c4185baac1
Fix parserTests.js --filter

In node 0.10, we get a bunch of "recursive invocation of process.nextTick, use
setImmediate" errors on console when we use --filter.  Switch to setImmediate
to make node happy, and add a compatibility thunk in core-upgrade.js so that
we still work fine on node 0.8 as well.

Change-Id: I55679b2144bf71e6093bd77507e426ce4250989a