Updated mediawiki/extensions Project: mediawiki/extensions/DataTypes  a31d82f2f008f3b4e3b227b9c778eca0713ee2ac
Update DataType definition to match what we currently want it to be

Includes some minor cleanup and deprecation of bad code

This is a breaking change that requires the following commit
to be merged in Wikibase: https://gerrit.wikimedia.org/r/#/c/85528/

This change only updates the PHP definition. The JS one should
still be updated in the near future.

Change-Id: Ifad7cafdf2151e1948e51ed5555eecd45197d095