Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  f52136eada9db462ff4781a44480cdcdcff97ff9
Handle invalid keys in Lua-to-PHP calls for LuaStandalone

PHP can't handle having arrays/objects or functions as keys in its
arrays, so make sure we don't try to pass them from Lua. Booleans aren't
really well-handled either, so let's disallow them too.

Also, add tests for proper stringification of floats and infinities when
those are used as keys.

Note this behavior change is needed to match the change in LuaSandbox
for fixing bug 54527, but isn't itself a security issue.

Change-Id: I1e2951bbe8cb78358650ad377bf7119fcac4485d