Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  f8eb557a6416c155148e2a723b84fbfd6bf7d0dc
Account for absence of template description in #getHtml

The description property is optional and is explicitly set to
null by the parser in normalisation if there is none.

The HTML rendering of parameter descriptions (as oppposed to the
description of the template as a whole) already covered for this.
Used the same logic for the template parameter.

Also fixed:
* Localize the text "no description"
* Use !== null instead of isset() since the property is always
  set, we just need to know whether it is null or an object.

Bug: 54422
Change-Id: I86a40dbd1225feb54123e77a856b4ad6d525461b