Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  3e78911815599f8841c5f0c72dc44de0e64f9b5f
PaymentMethod utility class

The immediate gain is that utm_source payment method will be
constrained to sane values.  The payment method component will be
the name of the payment submethod "group", with an "r" prepended
for recurring donations.

There are also two long-term goals which are supported by this patch:
* Selecting the gateway by desired payment method
* Unique and meaningful designation for each payment method

Change-Id: I34602faa9d8cca99132aa3b8067247e0f0a58b9e