Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  87e680ae966efdd1be65cc0065a2edc89042dfec
Update EasyRdf for for serialization fix.

The EasyRdf submodule was updated to a new version to avoid
a bug with the serialization of integers in turtle format.

This change includes regression tests which will fail if the
submodule is not updated.

IMPORTANT: this requires I3825b83d35b

IMPORTANT: when merging this, send an email to wikidata-tech
telling people to run `git submodule update` after pulling master.
Otherwise, RDF tests will fail on master.

Change-Id: I6f835fb8d007a0e9c5795b1ff30d969e977b4b37