Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  a622a7e6878ea661390a52f7cdc031c51f15cf5a
Add caching & cache invalidation for Campaign pages

Cache into memcached on first parse. Also calculate
the list of templates used in the campaign JSON,
and insert that into templatelinks. Then use
the JobQueue and LinksUpdate to invalidate the
cache on the Campaign whenever any of the templates
it depends on is changed.

Store an invalidation timestamp in a separate key,
and consider all cache items as valid only if they
have a timestamp value greater than or equal to the
invalidation timestamp

Change-Id: I79fb24997c20118e229585a4c97b45a173f466b1