Updated mediawiki/extensions Project: mediawiki/extensions/Flow  2837c75431adb89170a4ac48f93a30f64de18764
Flow: Add Mention notification.

Currently requires a double-parsing of wikitext to HTML because we
no longer use Parsoid to convert to HTML.

Also includes a refactoring of the notification code, attempting to
put it in its own self contained module. Unfortunately, this means
that it needs to be passed all the way down the tree from
WorkflowLoader in order to be used.

Now also includes a general improvement of notification code, such
as the addition of primary and secondary links, emailing and
bundling.

Dependent on talk page occupation because I've pre-adjusted it to
be compatible with my refactoring of UrlGenerator

Change-Id: Ic6bcd10cfd2cb96a5e0b13f814628fcc6df5d9ef