Updated mediawiki/extensions Project: mediawiki/extensions/PGFTikZ  be2e83e33d32ba9655be08c004a383003de8fac2
PGFTikZ extension: initial import

Tag extension (<PGFTIkZ> tag) used to input and render PGF/TikZ code.
The code is placed in a temporary LaTeX file, compiled using the
following sequence of commands:
-latex (generates a .dvi from the generate .tex file)
-dvips (generates a .eps file from the .dvi file)
-epstool (extracts a minimal bounding box from the .eps file)
-convert (ImageMagick converts to the final file type)
The generated image is uploaded to the wiki, and the source used to
generate it is saved in the image's wikipage.  When parsing the tag,
the source code is tested against the saved code to skip compilation
if it is not needed (see mediawiki.org/wiki/Extension:PGFTikZ).
Update after review.

Change-Id: I261a697d687c8bde84dfd4d3d5c3c28962ea6b65