Updated mediawiki/extensions Project: mediawiki/extensions/DataValues  de6dfbf0f3ed3d22714e63efc86110e7767880b0
(hotfix) Make inherit work with IE 8 again

(bug 53533)
In IE < 9, explicitly overwritten "constructor" properties on plain objects are still set as not
enumerable. So jQuery.extend won't copy the "constructor" value of given objects to its target
object.
In the inherit function we did use
  jQuery.extend( new NewPrototype(), { constructor: NewConstructor }, members );
and assumed that the resulting object would hold "NewConstructor" as "constructor" which would not
be the case in IE < 9 as explained above.
The ultimate result of this was that this.constructor has been set to wrong values in certain
instances in IE < 9.

Change-Id: I2710e0cea03cecfb50fcb28bbe0c19c9c8c4c64d