Updated mediawiki/extensions Project: mediawiki/extensions/Math  679ce7fe2fedc3790098aedb059646a66689c17e
Log texvc errors in 'texvc' log channel

This patch adds wfDebugLog() calls for conditions related to texvc invocations.
The logs are grouped under a 'texvc' group, added for this purpose. The reason
for logging into a separate channel is that the Math channel is too verbose for
production use on the Wikimedia cluster.

Change-Id: I05a17a0230f49f5d698b91617d06b3e3f838b67d