Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  72f36962564a12cebfb0d810ebdc720e5aa32fcc
corrected issue where batch upload breaks off early

the main issue was that while the metadata batch job was creating media file batch
jobs it would first evaluate the media file URL and then use the final file name it found
to create a media file batch job title. this worked as long as the media file URL was valid.
if there was a problem with the media file URL it would error out the metadata batch job
and stop it from creating any remaining media file batch jobs.

the solution implemented here was to use a generic media file batch job title and let
the media file batch job evaluate the media file URL when it runs. this isolates the
metadata batch job from any media file URL errors, so that it can continue to create
media file batch jobs, and only the individual media file batch job will error out if
there is something wrong with its media file URL.

another issue that was identified, while working on this main issue, had to do with
keeping track of the record count. record-count now always refers to the total
metadata records found and record-current is the current record the application
is processing.

bryan davis brought up the issue of unique batch job title names. after a short
discussion we decided that adding a uniqid to the title would take care of
the issue.

* GWToolset.i18n.php
  - added key 'gwtoolset-no-url-to-evaluate'
  - changed key 'gwtoolset-dom-record-issue'

* includes/Config.php
  - changed comments from {boolean} to {bool} for consistency

* includes/Forms/PreviewForm.php
  - changed $user_options['record-begin'] value from $user_options['record-count']
    to $user_options['record-current']

* includes/Handlers/Forms/MetadataMappingHandler.php
  - added line breaks to code that is > 100 characters
  - added record-current variable
  - changed set-up of record-begin variable
  - changed set-up of record-count variable
  - re-factored processMatchingElement()
  - set $_POST['record-begin'] to $user_options['record-current'] instead of
    $user_options['record-count']

* includes/Handlers/UploadHandler.php
  - added exception when $url provided to evaluateMediafileUrl() is empty
  - re-factored getTitle()
  - re-factored saveMediafileAsContent()
  - re-factored saveMediafileViaJob()
  - removed saveMediaFile()

* includes/Handlers/Xml/XmlMappingHandler.php
  - changed use of $user_options['record-count'] to $user_options['record-current']

* includes/Jobs/UploadMediafileJob.php
  - added use statements
  - changed use of error_log() to $this->setLastError()
  - expanded validateParams() to accommodate new params
  - passing $this->params['user_options'] instead of $this->params to
    saveMediafileAsContent();
  - re-factored processMetadata()
  - re-located validateParams() within the class
  - removed class variables

* includes/Jobs/UploadMetadataJob.php
  - created a processMetadata() method to match similar use in
    includes/Jobs/UploadMediafileJob.php

* re: #179/assembla - Batch upload breaks off early

Change-Id: I595168caba21fa23bda202724933f38de96f322a