Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  f81de3e1e9025751f6a7abf2934e267594c01c88
Initialize the save dialog correctly even if the sanity check failed

A very weird if statement caused pretty much all logic for displaying
the 'save' panel to be skipped if the sanity check had failed. The
result was that all buttons were displayed at the bottom of the
save dialog, even those that don't make sense in that context.

Change-Id: I628ebc05830fb25d67ff181852a1e56f0e25dee9