Updated mediawiki/extensions Project: mediawiki/extensions/Flow  cefdb478dc46956a04bf31f41b60972fd4d0d20c
Refactor information collection for i18n to dedicated class

Attempts to refactor many of the longer and multi-line php statements
out of post.html.php and into a helper class.  For the most part these
are i18n messages rendering.

The primary goal of moving the i18n messages from post.html.php into
View\Post is that I think they are a primary source of the 'php mess'
that is infiltrating the html markup we want to focus on in template
files.  The various i18n messages need to collect information from
the data models before they can output the information, so this patch
moves that information collection into a class dedicated to that
aspect of rendering.

Change-Id: I6024f66becb8b858b824f9e9f5ca04e07d0f91f3