Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  7db65f386c91fdfc4d05a309964295aa8068e34a
Rename @emits to @fires so we're forward compatible with JSDuck 5

Instead of using @emits in both, use our custom @fires in
production (JSDuck 4), and in the future it'll just naturally
use the native one.

This way we can also index oojs without issues, which seems to
have started using @fires already.

Change-Id: I7c3b56dd112626d57fa87ab995d205fb782a0149