Updated mediawiki/extensions Project: mediawiki/extensions/Echo  e4f97802db3e319e75c91d3af4a5f920aeacaa63
Split user rights changes into their own category

This will allow users to disable them if they wish to.
By default, both web and email are enabled for all users, which
is the current default.

This also fixes a bug in the EventLogging portion where user rights
changes would be logged with a user id of -1, instead of the
actual user.

Bug: 55337
Change-Id: I026b9392b085e295fbcd375095c9aed682fc29a2