Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  d548405b73282330b42811f0d0753b69c07f4c6e
Rename --rebuild to --startOver

--rebuild was confusing because it wasn't the kind of reindex we ever
wanted to run in production.  --startOver carries the right connotations
for what it actually does: it is useful when trying to get the thing
working for the first time but it removes all the data that exists so
shouldn't be used if you don't want a substantial interruption in
service.

Bug: 56164
Change-Id: I6aede782e6a3da5b729d6ea8cd82411fed319837