Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5a4f58434cacd1bc0189757e56975f5030b37ff7
Fix ByPropertyListUnserializer tests

Previous to this the dataprovider for valid data
was not actually being used by any tests!

This change removes the pointless base class
which was only used by this test class as
well as adding a test method that actually
tests the data provided.

We also fix the expected output as expecting
a Snaks object is wrong

Change-Id: Ifaf92d0d61b4c4f3be6b290ba72ed07799854013