Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  0a869d365f281a0c983667df7e75b904d6217d99
re-work exception handling

Gergő Tisza was concenred about the fact that the extension was catching MWExceptions.
the solution we agreed upon was to use an extension specific exception object for user
errors and another for system/non-user correctable errors, which has been implemented
in this change-set.

a few files also include php doc clean-up, shortening of lines > 100 characters, and
removal of a few functions no longer used by the extension.

Change-Id: Ie9493be07ecb74cf382bf3561e1c036af6538ac2