Updated mediawiki/extensions Project: mediawiki/extensions/OpenStackManager  710f76ba4b7186cdb2f55594c94856a28b3ba257
Add a gui for creation/deletion of labs proxies.

This communicates with the 'Invisible Unicorn' API in front of
yuviproxy.

This patch introduces two new config settings, defining the
region and IP, respectively, of the proxy gateway.

$wgOpenStackManagerProxyServiceRegion = '';	69
$wgOpenStackManagerProxyGateway = '';	70

To make this code active, an endpoint for IU needs to be
configured in keystone with service name 'proxy'.

Change-Id: I9844258ef4fcf1045e02202fd3fe67d14a6ffad5