Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  dfc40c8f59ca428c9c672a8c8ee9d6f36076a6c0
Refactor the css for interlanguage links and autonym class

Move it to a seperate css file ext.uls.webfonts.css so that the RL
modules are more granular and clear.

Also get the @font-face css for Autonym font get ready on page load
so that the font applied automatically to any future elements on page.

Change-Id: Icb76df5baedaba8ab613d665a1be923779a6b9e7