Updated mediawiki/extensions Project: mediawiki/extensions/Translate  003a9cbac5bfb86954311a4596810d55ac3db721
Do not transform message of TPException as work around for php bug

Due to a complicated tangle of various events coming together that
triggers an obscure PHP bug, calling ->text() on the message
causes the Parser to throw extensions related to invalid marker.
See the bug for details.

This should really be fixed properly somehow in the Parser, but
I'm not quite sure how to do that cleanly. In any case, it doesn't
hurt to fix it here too. Even once this is fixed in core, we
would want this extension to not have the bug if used with
old versions of MediaWiki

This would only change things (afaict) if a TPException is
thrown which is uncaught (usually they're caught). Furthermore, it
really shouldn't affect anything as it's unlikely those messages
would have {{VARIABLES}} to expand.

Bug: 56226
Change-Id: Ifd16c7c8ec6753ea4c74e6a9100e9d147ca915e9