Updated mediawiki/extensions Project: mediawiki/extensions/Flow  dfa2466cffe4de767ac2a37c41cd692083ba03b6
Require core edit permission to perform write actions

To detect if the core 'edit' permission was required we were checking
the button type and restricting all actions with 'button-type' !== 'GET'.
This turns out to be invalid because some actions, like edit-title,
draw 'GET' buttons because additional user interaction is required. In
this case the edit-title page will have its own POST submit, but that
configuration isn't a part of the $wgFlowActions.

The new method to detect need of the core 'edit' permission is to be 
explicit.  Each entry in the $wgFlowActions array must now contain the
key 'performs-write' set to a boolean.  Any action who's result is to
write to the wiki must set this to true so the edit permission is
respected.

No change is made to the 'button-type' key of $wgFlowActions, it still
refers to the type of button output in post.html.php

Change-Id: I1c4445305483e7b3bdf8ab455d23abc3b924b43d