Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  504a57031666610ad69cfca665891dff211212e3
filebackend

i ran into an issue while looking into developing a chunked upload solution for uploading
xml metadata. the plan was to use the mw api to upload the metadata in slices and keep
them in the upload stash for use, and dispose of them when no longer needed.

after i sorted out the initial functionality, the mw api came back with a message that
xml file uploads were not allowed. this could be overcome by altering LocalSettings.php,
but i brought up the issue in irc with aaron schulz and chris steipp. chris said that
we should not be uploading xml data this way.

chris and aaron discussed a few ideas and came back with a few concepts:

1. download metadata from an external url.
   a. js could be used to poll the download and then allow continuation with the upload
      process once the metadata file has been received by the wiki.
   b. i mentioned that i would like to keep the form upload for this step and implement
      the download from url method in phase 2 or later; aaron was okay with this.

2. store the metadata in a private file backend.
   a. SwiftFileBackend on mediawiki servers.
   b. FSFileBackend locally or for wikis that do not have alternative FileBackend stores

3. cleanup the backend when finished with the metadata file.
4. make sure the documentation mentions the need to secure the file backend directory.

this changeset address this issue.

Change-Id: I5b7426bcfceec67909b81d561cad5e4526e51520