Updated mediawiki/extensions Project: mediawiki/extensions/CommonsMetadata  50aec033297e0abb5f792dbf4e874f09918490e3
Add generic CC license parsing

All the possible combinations of license type / version / country
was getting awkward to handle by a fixed whitelist, so it is now
replaced by splitting to license name to parts, and trying to
recognize them.

The disadvantage is that some invalid license names are accepted
(like CC-SA-BY-1.0 or CC-BY-SA-3.0-XX) - this should be of no
practical consequence.

Change-Id: I85af37607db3752b58e577d7ba2a53e91c49113a