Updated mediawiki/extensions Project: mediawiki/extensions/VectorBeta  0cfd5a68fbbbd65c7d6946bb102ef7dc5ef4d78b
Add beta preferences icons for Hebrew and Arabic

This patch reverses the beta preference icon, similarly to what
is done in other Beta features. However, because the image here
includes a letter, it needs to be adapted to other alphabets
and not just reversed, as it doesn't make sense to write an A
in the beginning of a Hebrew or an Arabic text.

This patch adds images for Arabic and Hebrew alphabets. The code
for picking the image is simple. If there's demand for more alphabets,
it can be made smarter.

Change-Id: I8a85689b8daa2ed805044925267cfeed956d3471