Updated mediawiki/extensions Project: mediawiki/extensions/Flow  611d34e7f16b5b2c369f05a7187058b8d6731721
Topic history has different cache key for loading and saving

Action from post editing/moderation would not reflect in the history till cache flush,
this is mainly due to some constraint in Index::cacheKey().  It wouldn't convert a
binary string to hex unless the key has an _id suffix.  In Topic block, we are passing
in UUID object, which would generate a cache key with hex value.  In insert hook, only
the binary string is passed and the key doesn't have _id suffix, which would generate a
cache key with binary value.

Bug: 56688

Change-Id: Iea3aff55b15a22adc14f1fd6c5044bc7901896fa