Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  1f5d274e8fa9bc21baad9f24f86d380a04697248
whitelisted-post

aaron schulz mentioned a concern about how the extension was storing the full $_POST in the
batch jobs. he requested that we, “take the fields that matter and do any sanity filtering needed”.

while working on this issue, i noticed that there is potential for a form field name clash between
the extension’s form fields and the form fields that are dynamically created based on the
mediawiki template chosen. in order to avoid this conflist, i “namespaced” the extension’s form
fields by prefixing them all with “gwtoolset-”.

this change-set addresses these issues.

* have also shortend code lines > 100 characters

Change-Id: I4dd6514b8d8f54f1efcc13aa88dc50cabe0b9977