Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  f496085017d9012709f41b2a28e98c25bb7bf3dd
Account for sanitization changes properly in InputWidget

setValue() doesn't do anything if this.value === value, but with
sanitization it's possible for that to be true while the value in
the DOM is out of sync and needs to be changed.

The fix is to check for this.value changing and the DOM changing
separately.

Change-Id: I5f571445f5729f5477902c155a4ee9588b7194a8