Updated mediawiki/extensions Project: mediawiki/extensions/ZeroRatedMobileAccess  66ce83b85f082f4097ff2150173a59e44f8ca57c
Changed warning message for Zero

In order to make a full page describing zero-rated content and the fact
that the user is about to leave the free zone, we need a much better,
easier to read and understand page with Accept & Go Back buttons.

I want to get this text in so that i18n team can localize it before
we push it to production. Also added "Don't ask again" text to be translated.

Change-Id: I2f227f42a6d0164ffa41ca862d723336143839c8