Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  ebaf2811361dace4154de2c59354cb5b333c986c
sanitized-exceptions

chris steipp requested that:

* in includes/Specials/SpecialGWToolset.php, line 143: Please log, don't print_r

chris steipp suggested that:

* it would make things easier if you could escape the exception message, or
  demonstrate that it's been properly sanitized.

i augmented the GWTException class so that it now takes in wfMessage() keys and their
parameters and uses its ->parse() method to sanitize the message or assumes a regular
$message string and sanitizes it before use. adjusted all throw GWTExceptions in order
to properly use the new functionality.

i also corrected two bugs i found while working on this issue: the re-upload media file
checkbox wasn’t being picked-up and when no metadata file was provided for upload, a
MWException was being thrown instead of providing a user message.

Change-Id: Ic5626703262e5782af90d30a3c9aff35a898f35c