Updated mediawiki/extensions Project: mediawiki/extensions/Flow  b6cfc1c83b378ccf52eb622d8ff861a7cf6d1082
Change suppression terminology in code

Well, we don't *really* need this...
There is some confusion about the term "censor" still living in i18n msg keys
that needs to be addressed:
https://translatewiki.net/wiki/Thread:Support/Term_%22censor%22_in_Flow

I'd rather have all the code use the same terminology to avoid confusion.

If ever, I think now is the time to do this: we're not yet live anywhere, so the
change is cheap.

Change-Id: I1cffe3d4b561787580b5b450190ef54559c8347e