Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  dc98229d336be52ea8f28874a6c3001eac36e42f
server: make master and worker exit(0)

For some reason the master and workers process are exiting with error
code 1.  Seems to me that is the normal behavior and thus it deserves an
exit code of 0.

That would let us kill the server and keep proceeding in Jenkins (which
is usually set to abort whenever a command exit code is not 0).

Change-Id: I25177f911ed54a488579958d0a6236ef36aa2597

server: list number of worker about to be spawned

I am too lazy to count the worker(###) loading... lines so make the
server dump the number of workers it is about to spawn.

Change-Id: I3045e8b7a3ac7de270eb856d8fb6c04e1e6f6e86