Updated mediawiki/extensions Project: mediawiki/extensions/Comments  8440d877cbce0eea8b31e72d1b16b6c9995fdc5b
{{NUMBEROFCOMMENTSPAGE}} magic word/parser funct

Magic word to allow getting the number of comments on the current page,
and number of comments on any given page ({{NUMBEROFCOMMENTS:<page}}). To allow both situations, a magic word and parser function are needed. (This is what makes the code a bit messy). To organise things, I have moved the existing
{{NUMBEROFCOMMENTS}} magic word code into NumberOfComments.php, along with
the new magic word/parser function code, some of which has been integrated
into the existing {{NUMBEROFCOMMENTS}} hooks.

Extension version bumped up to 2.9

Change-Id: Ibc66c0d1e26653136f59b72be6fa17650ac740a8