Updated mediawiki/extensions Project: mediawiki/extensions/CryoKey  7b8c8d48db4df41a21dec3bdac0551c0f73c5430
Support new centralized CryoKey Lite API

Code from the original CryoKey.js is now available on the main CryoKey
page. Therefore, to keep things up to date, we refer to that JS file for
core functionality. Also, we make use of the new JSON redemption response
to future-proof the extension.

Change-Id: I22ef646953ef4346a498ab5382cc66ba6dd14046