Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  ccab415701f3585aa0c28c17fe564798ba56b531
Don't cache volatile wikitext

Certain wikitext, such as that containing Cite.php <ref> or <references>
or the #tag versions of the same, should not be cached. This uses the
isVolatile method added to PPFrame in I95b3cf87 to avoid caching the
preprocessed output of such wikitext from frame:preprocess and similar
methods.

Bug: 46815
Change-Id: I1084f87fd863eb22f2f3f3d3ff308b24e20a08ef